Add files via upload
This project controls and supply water to the plants according to their requirement , like different plants need different amount of water. The main application of this project is to control the flow of water based on the moisture of the soil and gives real time surveillance to the owners who stay far away from the farms. Wastage of water is reduced in 
this system. This is achieved by using a smart platform of IoT and solenoid valves to control the flow of water based on the moisture of the soil . This project also allows surveillance on the personnel and their crops so as to not occur losses. At last this system help to reduce soil erosion and nutrient leaching. This project is feasible and cost-effective for optimizing water resources for agricultural production.